+++ crates/behavior/src/actor/mod.rs
@@ -6,10 +6,10 @@ pub use addressing::{
     EstablishedRecipient, InterpretEstablished, MailAddr, Recipient,
 };
 pub use creation::{
-    AllocationRejection, BirthMode, BirthNodeProtocols, BirthProtocol, BirthProtocolAt,
-    BirthProtocolHead, BirthProtocolProduct, BirthProtocolTail, BirthProtocols, Births,
-    ChildChoice, ChildCons, ChildDelivery, ChildHead, ChildPosition, ChildProduct, ChildRole,
-    ChildRoute, ChildTail, Children, ChildrenError, Create, CreationKind, CreationRejection,
-    DispatchBirth, DispatchBirthAt, EstablishedCreation, InstallBirth, NoBirthProtocols, NoBirths,
-    NoChildren, RoleChild, RoleProtocol,
+    AllocationRejection, BirthMode, BirthNodeMapper, BirthNodeProtocols, BirthProtocol,
+    BirthProtocolAt, BirthProtocolHead, BirthProtocolProduct, BirthProtocolTail, BirthProtocols,
+    Births, ChildChoice, ChildCons, ChildDelivery, ChildHead, ChildPosition, ChildProduct,
+    ChildRole, ChildRoute, ChildTail, Children, ChildrenError, Create, CreationKind,
+    CreationRejection, DispatchBirth, DispatchBirthAt, EstablishedCreation, FoldBirthNode,
+    FoldedBirthNode, InstallBirth, NoBirthProtocols, NoBirths, NoChildren, RoleChild, RoleProtocol,
 };